+++ BoostClusteringMaB/BoostClusteringMaB/Clustering/ClusteringProtocol.swift
@@ -1,0 +1,19 @@
+//
+//  ClusteringProtocol.swift
+//  BoostClusteringMaB
+//
+//  Created by ParkJaeHyun on 2020/12/02.
+//
+
+import UIKit
+
+protocol ClusteringTool: class {
+    func convertLatLngToPoint(latLng: LatLng) -> CGPoint
+}
+
+protocol ClusteringData: class {
+    func redrawMap(_ latLngs: [LatLng],
+                   _ pointSizes: [Int],
+                   _ bounds: [(southWest: LatLng, northEast: LatLng)],
+                   _ convexHulls: [[LatLng]])
+}
