+++ BoostClusteringMaB/BoostClusteringMaB/CoreData/Models/ManagedPOI+CoreDataClass.swift
