+++ BoostClusteringMaB/BoostClusteringMaB/CoreData/Models/ManagedPOI+CoreDataProperties.swift
