+++ BoostClusteringMaB/BoostClusteringMaB/Models/POIModel.swift
