+++ BoostClusteringMaB/BoostClusteringMaB/POIRawData/json/gangnam_8000.json
