+++ BoostClusteringMaB/BoostClusteringMaB/Scenes/Detail/DetailCollectionViewCell.swift
@@ -11,7 +11,11 @@ class DetailCollectionViewCell: UICollectionViewCell {
     @IBOutlet var nameLabel: UILabel!
     @IBOutlet var categoryLabel: UILabel!
     @IBOutlet var storeImageView: UIImageView!
-    
+
+    override func prepareForReuse() {
+        storeImageView.image = UIImage(named: "icon")
+    }
+
     func configure(poi: ManagedPOI) {
         nameLabel.text = poi.name
         categoryLabel.text = poi.category
