+++ BoostClusteringMaB/BoostClusteringMaB/Scenes/MainView/MainPresenter.swift
@@ -25,9 +25,7 @@ final class MainPresenter: MainPresentationLogic, ClusteringData {
             NMGLatLngBounds(southWest: NMGLatLng(lat: $0.lat, lng: $0.lng),
                             northEast: NMGLatLng(lat: $1.lat, lng: $1.lng))
         }
-        let count = pointSizes.reduce(0){
-            $0 + $1
-        }
+        let count = pointSizes.reduce(0) { $0 + $1 }
         let newPolygons = NMFPolygonOverlay.polygonOverlays(convexHulls: convexHulls)
         let newViewModel = ViewModel(markers: newMarkers, polygons: newPolygons, bounds: newBounds, count: count)
